Modernized code, set it to be mypy compliant
