+++ sample.env
@@ -1,0 +1,1 @@
+TOKEN=INSERT_TOKEN_HERE
