+++ telibot/core/custom_exceptions.py
@@ -2,7 +2,9 @@ class TooLongException(Exception):
     def __init__(self, duration, max_duration: int = 6):
         self.duration = duration
         self.max_duration = max_duration
-        super().__init__(f"A duração do audio: {duration} excede o máximo pertimido de: {max_duration} segundos.")
+        super().__init__(
+            f"A duração do audio: {duration} excede o máximo pertimido de: {max_duration} segundos."
+        )
 
 
 class NotAudioException(Exception):
@@ -15,4 +17,6 @@ class FileSizeException(Exception):
     def __init__(self, size: int, max_size: int):
         self.size = size
         self.max_size = max_size
-        super().__init__(f"O arquivo tem tamanho: {size} bytes, excede o máximo permitido de: {max_size} bytes.")
+        super().__init__(
+            f"O arquivo tem tamanho: {size} bytes, excede o máximo permitido de: {max_size} bytes."
+        )
